refactor(queue): keep late batch contract internal
